Refactor the filter reducer
